+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 Features:
 - [alerts] alerts table default order should be by creation time newest at the top
 - [core] allow tracking Countly dashboard usage with Countly
+- [sdk] Improved and added new Server Config options
 
 Enterprise Features:
 - [journey_engine] Editing/Deleting/Duplication of blocks and version management